data_gen_server/pyproject.toml
[project]
name = "data_gen_server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.0",
    "starfish-core>=0.1.2",
    "datasets>=3.6.0",
    "openai-agents>=0.0.15",
]